Update the loan repayment - Interest
The current module calculates the payment schedule, without taking into account the actual number of days for each term. The interest to repay for different terms may differ from the actual bank amount. Therefore, it is considerable for user to update the amounts of repayment including both the capital and interest.